Remove use of make; fold profile into unittests
Since setuptools can handle installing scripts, we just install the
scripts via a specification in setup.py.  This allows us to simplify
the entire package, and ready for making this generally available via
PyPI.

We also simplify the unittests by merging the profile script directly
into the unittests script.

Finally, we drop the need to maintain executable scripts for Python 2
and Python 3. If the environment is Python 2, or its Python 3, the
script still works.
