+++ unittests
@@ -7,12 +7,34 @@ mkdir -p $tmp
 
 # trap "rm -rf $tmp" EXIT INT QUIT
 
+python setup.py build
+
+# So *always* use the in-tree version of configtools here.
+# prefix better be right at this point
+if [[ ! -f build/scripts-2.7/getconf.py ]]
+then
+    echo "Cannot find getconf.py: build/scripts-2.7/getconf.py" 1>&2
+    exit 1
+fi
+
+ppathins() {
+    if [ -d "$1" ] && [[ ":$PYTHONPATH:" != *":$1:"* ]]; then
+        PYTHONPATH="$1:$PYTHONPATH"
+    fi
+}
+
+export PYTHONPATH
+ppathins $PWD/build/lib
+
+bin=$PWD/build/scripts-2.7
+
 function run_test {
     testnum=$1
     shift
     dir=$1
+    shift
 
-    eval "$*" > $tmp/test.out 2>&1
+    eval "$bin/$dir $*" > $tmp/test.out 2>&1
 
     echo Test $testnum
     if diff $tmp/test.out gold/$dir/$testnum.out ; then
@@ -21,9 +43,6 @@ function run_test {
         echo FAIL
     fi
 }
-
-make
-. ./profile
 
 run_test 1 getconf.py --config samples/foo.conf bar foo
 run_test 2 getconf.py --config samples/foo.conf baz foo
